+++ src/main/java/ru/hh/resumebuilderbot/message/handler/MessageHandler.java
@@ -1,7 +1,7 @@
-package ru.hh.resumebuilderbot.message_handler;
+package ru.hh.resumebuilderbot.message.handler;
 
 import ru.hh.resumebuilderbot.Answer;
-import ru.hh.resumebuilderbot.question_generator.QuestionGenerator;
+import ru.hh.resumebuilderbot.question.generator.QuestionGenerator;
 
 public interface MessageHandler {
     public QuestionGenerator handle(Answer answer);
