+++ src/main/java/ru/hh/resumebuilderbot/message/handler/ShowMessageHandler.java
@@ -1,8 +1,8 @@
-package ru.hh.resumebuilderbot.message_handler;
+package ru.hh.resumebuilderbot.message.handler;
 
 import ru.hh.resumebuilderbot.Answer;
-import ru.hh.resumebuilderbot.question_generator.QuestionGenerator;
-import ru.hh.resumebuilderbot.question_generator.ShowAllQuestionGenerator;
+import ru.hh.resumebuilderbot.question.generator.QuestionGenerator;
+import ru.hh.resumebuilderbot.question.generator.ShowAllQuestionGenerator;
 
 public class ShowMessageHandler implements MessageHandler {
     @Override
