+++ src/main/java/ru/hh/resumebuilderbot/message/handler/UnknownMessageHandler.java
@@ -1,9 +1,9 @@
-package ru.hh.resumebuilderbot.message_handler;
+package ru.hh.resumebuilderbot.message.handler;
 
 
 import ru.hh.resumebuilderbot.Answer;
-import ru.hh.resumebuilderbot.question_generator.FixedQuestionGenerator;
-import ru.hh.resumebuilderbot.question_generator.QuestionGenerator;
+import ru.hh.resumebuilderbot.question.generator.FixedQuestionGenerator;
+import ru.hh.resumebuilderbot.question.generator.QuestionGenerator;
 import ru.hh.resumebuilderbot.TextsStorage;
 
 public class UnknownMessageHandler implements MessageHandler {
